+++ app/src/main/java/org/gdg/frisbee/android/utils/PlusUtils.java
@@ -1,5 +1,7 @@
 package org.gdg.frisbee.android.utils;
 
+import android.net.Uri;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
 import com.google.android.gms.common.api.GoogleApiClient;
@@ -20,7 +22,7 @@ public class PlusUtils {
         return plusPerson != null ? plusPerson.getId() : null;
     }
 
-    public static String createProfileUrl(String gplusId) {
-        return "https://plus.google.com/" + gplusId;
+    public static Uri createProfileUrl(@NonNull final String gplusId) {
+        return Uri.parse("https://plus.google.com/" + gplusId);
     }
 }
